refactor configvalidation package
